Updates from PR feedback
