+++ client/web/src/theme.test.tsx
@@ -2,8 +2,7 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
 import { renderHook, act } from '@testing-library/react'
 
-import { ThemePreference } from './stores/themeState'
-import { useThemeProps } from './theme'
+import { ThemePreference, useThemeProps } from './theme'
 
 // Don't test reacting to system-wide theme changes, for simplicity. This means that
 // observeSystemIsLightTheme's initial value will be used, but it will not monitor for subsequent
@@ -16,10 +15,6 @@ jest.mock('@sourcegraph/wildcard', () => {
         useObservable: () => undefined,
     }
 })
-
-// jest.mock('@sourcegraph/shared/src/settings/temporary/useTemporarySetting', () => ({
-//     useTemporarySetting: () => ['system', () => {}],
-// }))
 
 const mockSystemTheme = (systemTheme: 'light' | 'dark') => {
     window.matchMedia = query => {
@@ -42,6 +37,8 @@ describe('useTheme()', () => {
             expect(result.current.themePreference).toBe(ThemePreference.System)
             expect(document.documentElement.classList).toContain('theme-light')
             expect(document.documentElement.classList).not.toContain('theme-dark')
+            // Local storage item not set by default, will get set when the preference is changed
+            expect(localStorage).toHaveLength(0)
         })
 
         it('dark', () => {
@@ -53,6 +50,8 @@ describe('useTheme()', () => {
             expect(result.current.themePreference).toBe(ThemePreference.System)
             expect(document.documentElement.classList).toContain('theme-dark')
             expect(document.documentElement.classList).not.toContain('theme-light')
+            // Local storage item not set by default, will get set when the preference is changed
+            expect(localStorage).toHaveLength(0)
         })
     })
 
@@ -66,6 +65,9 @@ describe('useTheme()', () => {
             expect(result.current.themePreference).toBe(ThemePreference.Light)
             expect(document.documentElement.classList).toContain('theme-light')
             expect(document.documentElement.classList).not.toContain('theme-dark')
+            expect(JSON.parse(localStorage.getItem('temporarySettings') ?? '')).toEqual({
+                'user.themePreference': 'light',
+            })
         })
 
         it('dark', () => {
@@ -77,6 +79,9 @@ describe('useTheme()', () => {
             expect(result.current.themePreference).toBe(ThemePreference.Dark)
             expect(document.documentElement.classList).toContain('theme-dark')
             expect(document.documentElement.classList).not.toContain('theme-light')
+            expect(JSON.parse(localStorage.getItem('temporarySettings') ?? '')).toEqual({
+                'user.themePreference': 'dark',
+            })
         })
 
         it('system', () => {
@@ -88,6 +93,9 @@ describe('useTheme()', () => {
             expect(result.current.themePreference).toBe(ThemePreference.System)
             expect(document.documentElement.classList).toContain('theme-dark')
             expect(document.documentElement.classList).not.toContain('theme-light')
+            expect(JSON.parse(localStorage.getItem('temporarySettings') ?? '')).toEqual({
+                'user.themePreference': 'system',
+            })
         })
     })
 
@@ -105,6 +113,7 @@ describe('useTheme()', () => {
         expect(result.current.themePreference).toBe(ThemePreference.Dark)
         expect(document.documentElement.classList).toContain('theme-dark')
         expect(document.documentElement.classList).not.toContain('theme-light')
+        expect(JSON.parse(localStorage.getItem('temporarySettings') ?? '')).toEqual({ 'user.themePreference': 'dark' })
 
         // Change to system.
         act(() => {
@@ -114,5 +123,8 @@ describe('useTheme()', () => {
         expect(result.current.themePreference).toBe(ThemePreference.System)
         expect(document.documentElement.classList).toContain('theme-light')
         expect(document.documentElement.classList).not.toContain('theme-dark')
+        expect(JSON.parse(localStorage.getItem('temporarySettings') ?? '')).toEqual({
+            'user.themePreference': 'system',
+        })
     })
 })
